+++ packages/navi-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @awell-health/navi-js-react
 
+## 0.0.7
+
+### Patch Changes
+
+- Add deterministic session id creation from session token data
+- Updated dependencies
+  - @awell-health/navi-core@0.0.7
+  - @awell-health/navi-js@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes